Added morphism class. Refactoring: used size_t wherever possible

+++ include/DeterministicBuechiAutomaton.h
@@ -13,8 +13,8 @@ namespace omalg {
   public:
     DeterministicBuechiAutomaton (std::vector<std::string> theStateNames,
                                   std::vector<std::string> theAlphabet,
-                                  int theInitialState,
-                                  std::vector<std::vector<int> > theTransitionTable,
+                                  size_t theInitialState,
+                                  std::vector<std::vector<size_t> > theTransitionTable,
                                   std::vector<bool>  theFinalStates);
     OmegaSemigroup* toOmegaSemigroup() const override;
     std::string description() const override;
