+++ include/IOHandler.h
@@ -29,17 +29,17 @@ namespace omalg {
 	  void writeOmegaSemigroupToStream(OmegaSemigroup const &S, std::ostream &out);
 
 	  std::istream& safeGetLine(std::istream &in, std::string &str);
-	  void checkReadTillEnd(int lineNo, int lines);
-	  std::list<std::string> readNamesIntoList(std::vector<std::string> const &lines, int lineNo);
+	  void checkReadTillEnd(size_t lineNo, size_t lines);
+	  std::list<std::string> readNamesIntoList(std::vector<std::string> const &lines, size_t lineNo);
 
-	  std::vector<std::vector<std::set<int> > > buildTransitionRelation(std::list<std::string> const &transitions,
+	  std::vector<std::vector<std::set<size_t> > > buildTransitionRelation(std::list<std::string> const &transitions,
 	                                                                 std::vector<std::string> const &stateVector,
 	                                                                 std::vector<std::string> const &letterVector,
-	                                                                 int transNo);
-	  std::vector<std::vector<int> > buildTransitionTable(std::list<std::string> const &transitions,
+	                                                                 size_t transNo);
+	  std::vector<std::vector<size_t> > buildTransitionTable(std::list<std::string> const &transitions,
 	                                                                 std::vector<std::string> const &stateVector,
 	                                                                 std::vector<std::string> const &letterVector,
-	                                                                 int transNo);
+	                                                                 size_t transNo);
 		  
   public:
