+++ include/NondeterministicOmegaAutomaton.h
@@ -10,9 +10,9 @@ namespace omalg {
 
   class NondeterministicOmegaAutomaton: public virtual OmegaAutomaton {
 	private:
-		std::vector<std::vector<std::set<int> > > transitionRelation;
+		std::vector<std::vector<std::set<size_t> > > transitionRelation;
 	protected:
-		NondeterministicOmegaAutomaton(std::vector<std::vector<std::set<int> > > theTransitionRelation);
+		NondeterministicOmegaAutomaton(std::vector<std::vector<std::set<size_t> > > theTransitionRelation);
   public:
     virtual std::string description() const override;
   };
